+++ packages/platforms/browser/lib/config.ts
@@ -8,6 +8,7 @@ import {
 } from '@bugsnag/core-performance'
 import { defaultNetworkRequestCallback, isNetworkRequestCallback, type NetworkRequestCallback } from './network-request-callback'
 import { isRoutingProvider, type RoutingProvider } from './routing-provider'
+import { defaultSendPageAttributes, isSendPageAttributes, type SendPageAttributes } from './send-page-attributes'
 
 export interface BrowserSchema extends CoreSchema {
   autoInstrumentFullPageLoads: ConfigOption<boolean>
@@ -17,6 +18,7 @@ export interface BrowserSchema extends CoreSchema {
   routingProvider: ConfigOption<RoutingProvider>
   settleIgnoreUrls: ConfigOption<Array<string | RegExp>>
   networkRequestCallback: ConfigOption<NetworkRequestCallback>
+  sendPageAttributes: ConfigOption<SendPageAttributes>
 }
 
 export interface BrowserConfiguration extends Configuration {
@@ -27,6 +29,7 @@ export interface BrowserConfiguration extends Configuration {
   routingProvider?: RoutingProvider
   settleIgnoreUrls?: Array<string | RegExp>
   networkRequestCallback?: NetworkRequestCallback
+  sendPageAttributes?: SendPageAttributes
 }
 
 export function createSchema (hostname: string, defaultRoutingProvider: RoutingProvider): BrowserSchema {
@@ -70,6 +73,11 @@ export function createSchema (hostname: string, defaultRoutingProvider: RoutingP
       defaultValue: defaultNetworkRequestCallback,
       message: 'should be a function',
       validate: isNetworkRequestCallback
+    },
+    sendPageAttributes: {
+      defaultValue: defaultSendPageAttributes,
+      message: 'should be an object',
+      validate: isSendPageAttributes
     }
   }
 }